backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
numpy
google-genai
sqlalchemy
PyJWT
bcrypt
python-multipart
requests
# Rate limiting
slowapi
# PostgreSQL support (optional; only needed when DATABASE_URL points to PostgreSQL)
psycopg2-binary
